requirements.txt
aiohttp>=3.7.4.post0
async-timeout>=3.0.1
attrs>=21.2.0
beautifulsoup4>=4.9.3
certifi>=2020.12.5
chardet>=4.0.0
colorthief>=0.2.1
discord-components>=0.5.1.1
discord-ext-menus>=1.0.0a30+g309c702
discord.py>=1.7.1
emoji>=1.2.0
idna>=2.10
imageio>=2.9.0
multidict>=5.1.0
numpy>=1.20.3
opencv-python>=4.5.2.52
Pillow>=8.2.0
psutil>=5.8.0
pytz>=2021.1
requests>=2.25.1
soupsieve>=2.2.1
tabulate>=0.8.9
typing-extensions>=3.10.0.0
urllib3>=1.26.5
websockets>=9.0.2
yarl>=1.6.3
peewee>=3.14.4